gci: Consolidate build jobs in a single step
We were building the CLN version in the same environment that we
tested in, which is suboptimal. More importantly we were building the
same variants over and over again, wasting precious CI minutes. Now we
build in a preparatory step, put all variants into the cache, and then
test in a separate step instead.